+++ src/navigation.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { Navbar, Nav, NavItem } from 'react-bootstrap';
 import { Link } from 'react-router-dom';
 import { LinkContainer } from 'react-router-bootstrap';
+import { FormattedMessage } from 'react-intl';
 const isDev = require('electron-is-dev');
 
 const Navigation = () =>
@@ -23,19 +24,35 @@ const Navigation = () =>
             </LinkContainer>
           : null}
         <LinkContainer to="/schedule">
-          <NavItem eventKey={3}>Schedule</NavItem>
+          <NavItem eventKey={3}>
+            <FormattedMessage id={'nav.schedule'} defaultMessage={'Schedule'} />
+          </NavItem>
         </LinkContainer>
         <LinkContainer to="/records">
-          <NavItem eventKey={3}>Records</NavItem>
+          <NavItem eventKey={3}>
+            <FormattedMessage id={'nav.records'} defaultMessage={'Records'} />
+          </NavItem>
         </LinkContainer>
         <LinkContainer to="/results">
-          <NavItem eventKey={4}>Battle History</NavItem>
+          <NavItem eventKey={4}>
+            <FormattedMessage
+              id={'nav.results'}
+              defaultMessage={'Battle History'}
+            />
+          </NavItem>
         </LinkContainer>
         <LinkContainer to="/meta">
-          <NavItem eventKey={5}>League Meta</NavItem>
+          <NavItem eventKey={5}>
+            <FormattedMessage
+              id={'nav.league'}
+              defaultMessage={'League Meta'}
+            />
+          </NavItem>
         </LinkContainer>
         <LinkContainer to="/settings">
-          <NavItem eventKey={6}>Settings</NavItem>
+          <NavItem eventKey={6}>
+            <FormattedMessage id={'nav.settings'} defaultMessage={'Settings'} />
+          </NavItem>
         </LinkContainer>
       </Nav>
     </Navbar.Collapse>